Fix import statements and restructure code
