+++ bot/handlers/helpers/__init__.py
@@ -1,1 +1,6 @@
 from .generateSurahButtons import generateSurahButtons
+
+from .getAyahReply import getAyahReply
+from .getAyahButton import getAyahButton
+from .getValidReply import getValidReply
+from .getRandomAyah import getRandomAyah
